style: move todo row to another file
