+++ src/pages/index.tsx
@@ -2,7 +2,6 @@ import { Project } from '@prisma/client';
 import { Trash } from 'lucide-react';
 import { nanoid } from 'nanoid';
 import * as React from 'react';
-import CreatableSelect from 'react-select/creatable';
 import { useSubscribe } from 'replicache-react';
 
 import { useReplicache } from '@/hooks/useReplicache';
@@ -11,6 +10,7 @@ import { useSpace, useSpaceStore } from '@/hooks/useSpace';
 import Button from '@/components/buttons/Button';
 import Layout from '@/components/layout/Layout';
 import Seo from '@/components/Seo';
+import { TodoRow } from '@/components/todo/TodoRow';
 
 import { IDB_KEY } from '@/models/idb-key.model';
 import { TodoDetail } from '@/models/todo.model';
@@ -195,129 +195,5 @@ export default function HomePage() {
         </section>
       </main>
     </Layout>
-  );
-}
-
-export function TodoRow({
-  todo,
-  idbKey,
-}: {
-  todo: ConvertDate<TodoDetail>;
-  idbKey: string;
-}) {
-  const rep = useReplicache();
-  const spaceId = useSpace();
-  const [project, setProject] = React.useState(todo.projectId ?? '');
-  const [chosenLables, setChosenLables] = React.useState<
-    { label: string; value: string }[]
-  >([]);
-
-  React.useEffect(() => {
-    setChosenLables(
-      (todo.labelOnIssues ?? []).map((l) => ({
-        label: l.label.name.toLowerCase(),
-        value: l.label.name.toLowerCase(),
-      }))
-    );
-  }, [todo.labelOnIssues]);
-
-  const labels = [
-    { label: 'bug', value: 'bug' },
-    { label: 'feature', value: 'feature' },
-    { label: 'enhancement', value: 'enhancement' },
-    { label: 'documentation', value: 'documentation' },
-    { label: 'help wanted', value: 'help wanted' },
-    { label: 'good first issue', value: 'good first issue' },
-    { label: 'invalid', value: 'invalid' },
-    { label: 'question', value: 'question' },
-    { label: 'wontfix', value: 'wontfix' },
-  ];
-
-  const projects = useSubscribe(
-    rep,
-    async (tx) => {
-      const list = await tx
-        .scan<ConvertDate<Project>>({ prefix: IDB_KEY.PROJECT({ spaceId }) })
-        .entries()
-        .toArray();
-
-      // sort by title using localeCompare
-      list.sort((a, b) =>
-        a[1].name.localeCompare(b[1].name, undefined, {
-          numeric: true,
-        })
-      );
-      return list;
-    },
-    { default: [] }
-  );
-  return (
-    <div key={idbKey} className='flex items-center space-x-4'>
-      <button
-        onClick={() => {
-          rep?.mutate.todoDelete({
-            id: todo.id,
-          });
-        }}
-      >
-        <Trash size={15} />
-      </button>
-      <span>{todo.title}</span>
-      <select
-        name='project'
-        id='project'
-        value={project}
-        onChange={(e) => {
-          setProject(e.target.value);
-          rep?.mutate.todoUpdate({
-            id: todo.id,
-            projectId: e.target.value ?? null,
-          });
-        }}
-      >
-        <option value=''>None</option>
-        {projects.map(([idbKey, project]) => (
-          <option key={idbKey} value={project.id} className='text-orange-400'>
-            {project.name}
-          </option>
-        ))}
-      </select>
-      <span className='text-green-600'>#{todo.GithubIssue?.number}</span>
-      <CreatableSelect
-        isMulti
-        options={labels}
-        value={chosenLables}
-        onChange={(selected) => {
-          setChosenLables(Array.from(selected ?? []));
-        }}
-        closeMenuOnSelect={false}
-        onMenuClose={() => {
-          rep?.mutate.todoUpdate({
-            id: todo.id,
-            labelOnIssues: chosenLables.map((l) => {
-              // check is label exist
-              const existingLabel = todo.labelOnIssues?.find(
-                (loi) => loi.label.name.toLowerCase() === l.value
-              );
-              if (existingLabel) {
-                return existingLabel;
-              } else {
-                const labelId = nanoid();
-                return {
-                  id: nanoid(),
-                  label: {
-                    id: labelId,
-                    name: l.value,
-                    color: '#000000',
-                  },
-                  labelId,
-                  todoId: todo.id,
-                };
-              }
-            }),
-          });
-        }}
-      />
-    </div>
   );
 }
